Move add-cart functionality to detail page
