+++ Frontend/autozon/src/app/components/product-card/product-card.component.html
@@ -19,7 +19,4 @@
             Löschen
         </button>
     </mat-card-actions>
-    <mat-card-actions *ngIf="showCartButton()">
-        <button mat-raised-button color="primary" class="deleteButton" (click)="addToCart()">In den Warenkorb</button>
-    </mat-card-actions>
 </mat-card>
