+++ Frontend/autozon/src/app/pages/product-detail/product-detail.component.ts
@@ -1,6 +1,9 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { IProduct } from 'src/app/core/models/product.model';
 import { ProductIdService, ProductService } from 'src/app/core/services/product/product.service';
+import { IUser } from 'src/app/core/models/user.model';
+import { LoginService } from 'src/app/core/services/login/login.service';
+import { ICartModel } from 'src/app/core/models/cart.model';
 
 @Component({
   selector: 'app-product-detail',
@@ -9,8 +12,27 @@ import { ProductIdService, ProductService } from 'src/app/core/services/product/
 })
 export class ProductDetailComponent implements OnInit {
   product: IProduct = {};
-  constructor(private productIdService: ProductIdService, private productService: ProductService) {
-  }
+  user: IUser = {
+    id: 0,
+    lastName: '',
+    firstName: '',
+    companyName: '',
+    email: '',
+    phone: '',
+    streetAndHouseNumber: '',
+    zipCode: '',
+    city: '',
+    country: '',
+    role: '',
+    verified: false
+  };
+  route: string = window.location.pathname;
+
+  constructor(
+    private productIdService: ProductIdService,
+    private productService: ProductService,
+    private loginService: LoginService
+  ) {}
 
   ngOnInit(): void {
     const id = this.productIdService.getId();
@@ -20,6 +42,51 @@ export class ProductDetailComponent implements OnInit {
         this.product = res;
       }
     });
+    // Get currently authenticated user
+    this.loginService.getUser().subscribe(res => {
+      if (res !== undefined) {
+        this.user = res;
+      }
+    });
   }
 
+  /**
+   * Show cart button if authenticated user is customer and not on my-product page
+   */
+  showCartButton(): boolean {
+    if (this.user.role == 'customer' && this.route !== '/my-products') {
+      return true;
+    }
+    return false;
+  }
+
+  /**
+   * Add product to cart array in session storage
+   */
+  addToCart(): void {
+    if (this.product.id && this.user.role === "customer") {
+      const cartContent: string = sessionStorage.getItem('cart') || '[]';
+      var cartArray: ICartModel[] = JSON.parse(cartContent);
+      var found: boolean = false;
+
+      cartArray.forEach((entry: ICartModel) => {
+        // Increase amount if product is already in cart
+        if (entry.id === this.product.id) {
+          entry.amount+= 1;
+          found = true;
+        }
+      });
+
+      // Add new entry in cart array
+      if (!found) {
+        cartArray.push({
+          id: this.product.id,
+          amount: 1
+        });
+      }
+
+      // Save to sessions storage
+      sessionStorage.setItem('cart', JSON.stringify(cartArray));
+    }
+  }
 }
